Initial value for single select taxonomy
