+++ SharePointFramework/shared-library/src/components/CustomEditPanel/CustomEditPanelBody/FieldElements/TaxonomyFieldType.tsx
@@ -3,9 +3,12 @@ import { FieldContainer } from '../../../FieldContainer'
 import { useCustomEditPanelContext } from '../../context'
 import { FieldElementComponent } from './types'
 import { ModernTaxonomyPicker } from '@pnp/spfx-controls-react'
+import { Term, useInitialTaxonomyValues } from './useInitialTaxonomyValues'
 
 export const TaxonomyFieldType: FieldElementComponent = ({ field }) => {
   const context = useCustomEditPanelContext()
+  const mapInitialValues = useInitialTaxonomyValues()
+  const terms = context.model.get<Term[]>(field)
 
   return (
     <FieldContainer
@@ -17,6 +20,7 @@ export const TaxonomyFieldType: FieldElementComponent = ({ field }) => {
       <ModernTaxonomyPicker
         context={context.props.dataAdapter.spfxContext as any} // Newest version of the control requires this cast for now, as context type is incompatibale with other types of SPFxContext
         panelTitle={field.description || field.displayName}
+        initialValues={terms?.map(mapInitialValues)}
         label=''
         termSetId={field.getProperty('TermSetId')}
         onChange={(items) => context.model.set(field, items)}
